requirements.txt

# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
tensorflow==2.18.0
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
requests==2.31.0
python-dotenv==1.0.0
scikit-learn>=1.3.2
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv>=1.0.0
mediapipe==0.10.14
google-generativeai>=0.7.2

# RAG & AI Extensions
chromadb>=0.4.15
pypdf>=3.17.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-google-genai>=0.0.3
reportlab>=4.0.0


